Provides blanket implementation of DivisibleBy
Instead of implementing only for u64, we provide a blanket implementation
of DivisibleBy for any type that we can ask for the remainder of given
another instance of that type, and get back a value of that type. Which
includes, logically, anything that is a number in a fairly broad sense.